+++ wiki/Changelog.md
@@ -1,3 +1,6 @@
+Terminal features
+  * Linux console controls for underscore cursor size (mintty/wsltty#203).
+
 ### 3.1.0 (23 November 2019) ###
 
 Terminal features